fix(simulation): harden stop ABI and event buffer ownership

+++ engine/simulation/description/include/lux/engine/simulation/ScriptMountDescription.hpp
@@ -60,7 +60,7 @@ namespace lux::simulation
         STOP,
     };
 
-    enum class EBehaviorStopReason : std::uint8_t
+    enum class EBehaviorStopReason : std::uint32_t
     {
         MOUNT_REMOVED,
         ENTITY_DESTROYED,
